pyproject.toml
[tool.poetry]
name = "python-test-assigment"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [
    { include = "connectors", from = "src" },
    { include = "routes", from = "src" },
    { include = "utils", from = "src" }
]

[tool.poetry.dependencies]
python = "3.10.12"
fastapi-sessions = "^0.3.2"
fastapi = "^0.111.0"
python-dotenv = "^1.0.1"
google-api-python-client = "^2.137.0"
google-auth = "^2.32.0"
google-auth-oauthlib = "^1.2.1"
google-auth-httplib2 = "^0.2.0"
redis = "^5.0.7"
uvicorn = "^0.30.1"
openai = "^1.35.13"
langchain = "^0.2.7"
langchain-community = "^0.2.7"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
